refactor: deepen video application, scrape apply, and composite-run seams
Split video IPC into query/maintenance; consolidate scrape delivery into
videoScrapeApplyService; share composite field-group orchestration across hosts.
