+++ scripts/check-actress-boundaries.mjs
@@ -137,14 +137,20 @@ for (const specifier of importsOf(actressHandler)) {
   }
 }
 
+const videoApplicationSeams = new Set([
+  '../services/videoQueryService',
+  '../services/videoMaintenanceService'
+])
+
 const videoHandler = 'src/main/ipc/videoHandlers.ts'
 for (const specifier of importsOf(videoHandler)) {
   if (
     /\.\.\/(db|scrapers)\//.test(specifier) ||
-    (/\.\.\/services\//.test(specifier) &&
-      specifier !== '../services/videoApplicationService')
+    (/\.\.\/services\//.test(specifier) && !videoApplicationSeams.has(specifier))
   ) {
-    violations.push(`${videoHandler}: video IPC must delegate through its application service`)
+    violations.push(
+      `${videoHandler}: video IPC must delegate through query / maintenance seams`
+    )
   }
 }
 
@@ -177,7 +183,8 @@ for (const file of sourceFiles('src/main/db')) {
 }
 
 for (const file of [
-  'src/main/services/videoService.ts',
+  'src/main/services/videoMaintenanceService.ts',
+  'src/main/services/videoScrapeApplyService.ts',
   'src/main/services/actressQueryService.ts',
   'src/main/services/actressMaintenanceService.ts',
   'src/main/services/actressGalleryService.ts',
