+++ src/main/db/videoRepo.ts
@@ -1,17 +1,11 @@
 import { getDb } from './database'
-import type { ActressGender } from '@shared/actressTypes'
 import type { Video, VideoFile, VideoAsset, VideoDetail, VideoQuery, VideoListResult, VideoEditInput } from '@shared/videoTypes'
-import type { ScrapeResult, ScrapedActress, VideoBatchScrapeFilter, VideoBatchScrapeStatus, VideoScrapeField, VideoScrapeUpdateMode, VideoRematchScope } from '@shared/videoScrapeTypes'
-import { ALL_VIDEO_SCRAPE_FIELDS } from '@shared/videoScrapeTypes'
+import type { VideoBatchScrapeFilter, VideoBatchScrapeStatus, VideoRematchScope } from '@shared/videoScrapeTypes'
 import { upsertActressFromScrape } from './actressRepo'
 import { actressOwnedNamePatternSearchSql } from './actressSearchSql'
 import { ensureTag, pruneTagIfUnused } from './tagRepo'
 import { ensureFacetEntries } from './facetRepo'
 import { collectVideoLibraryCleanupHints, runLibraryCleanup } from './libraryCleanup'
-export interface VideoImageAvailabilityFacts {
-  coverAvailable?: boolean
-  samplePathsAvailable?: boolean
-}
 
 export interface NewVideo {
   code: string
@@ -631,111 +625,6 @@ export function removeManualVideoTag(videoId: number, tagId: number): void {
   }
 }
 
-function scrapedCastGender(a: ScrapedActress): ActressGender {
-  return a.gender ?? 'female'
-}
-
-/** Remove cast links for one gender only (NULL gender is treated as female). */
-function removeVideoActressesByGender(videoId: number, gender: ActressGender): void {
-  const db = getDb()
-  if (gender === 'female') {
-    db.prepare(
-      `DELETE FROM video_actress
-       WHERE video_id = ?
-         AND actress_id IN (
-           SELECT id FROM actresses WHERE gender IS NULL OR gender = 'female'
-         )`
-    ).run(videoId)
-  } else {
-    db.prepare(
-      `DELETE FROM video_actress
-       WHERE video_id = ?
-         AND actress_id IN (SELECT id FROM actresses WHERE gender = 'male')`
-    ).run(videoId)
-  }
-}
-
-function linkScrapedCastByGender(
-  videoId: number,
-  cast: ScrapedActress[],
-  gender: ActressGender,
-  actressAvatars: Map<string, string | null>
-): void {
-  const db = getDb()
-  for (const a of cast) {
-    if (scrapedCastGender(a) !== gender) continue
-    const actressId = upsertActressFromScrape(
-      a.name,
-      actressAvatars.get(a.name) ?? null,
-      gender
-    )
-    db.prepare('INSERT OR IGNORE INTO video_actress (video_id, actress_id) VALUES (?, ?)').run(
-      videoId,
-      actressId
-    )
-  }
-}
-
-function isBlankText(value: string | null | undefined): boolean {
-  return value == null || value.trim() === ''
-}
-
-function countVideoCastByGender(videoId: number, gender: ActressGender): number {
-  const db = getDb()
-  const condition =
-    gender === 'female' ? "(a.gender = 'female' OR a.gender IS NULL)" : "a.gender = 'male'"
-  return (
-    db
-      .prepare(
-        `SELECT COUNT(*) AS n
-         FROM video_actress va
-         JOIN actresses a ON a.id = va.actress_id
-         WHERE va.video_id = ? AND ${condition}`
-      )
-      .get(videoId) as { n: number }
-  ).n
-}
-
-function countScrapedVideoTags(videoId: number): number {
-  const db = getDb()
-  return (
-    db
-      .prepare('SELECT COUNT(*) AS n FROM video_tag WHERE video_id = ? AND origin = ?')
-      .get(videoId, 'scraped') as { n: number }
-  ).n
-}
-
-function countVideoExternalIds(videoId: number, sourceName?: string): number {
-  const db = getDb()
-  const row = sourceName
-    ? db
-        .prepare(
-          "SELECT COUNT(*) AS n FROM video_external_ids WHERE video_id = ? AND source = ? AND url IS NOT NULL AND trim(url) != ''"
-        )
-        .get(videoId, sourceName)
-    : db
-        .prepare(
-          "SELECT COUNT(*) AS n FROM video_external_ids WHERE video_id = ? AND url IS NOT NULL AND trim(url) != ''"
-        )
-        .get(videoId)
-  return (row as { n: number }).n
-}
-
-function countVideoExternalStats(videoId: number, sourceName?: string): number {
-  const db = getDb()
-  const row = sourceName
-    ? db
-        .prepare(
-          'SELECT COUNT(*) AS n FROM video_external_stats WHERE video_id = ? AND source = ? AND rating_average IS NOT NULL'
-        )
-        .get(videoId, sourceName)
-    : db
-        .prepare(
-          'SELECT COUNT(*) AS n FROM video_external_stats WHERE video_id = ? AND rating_average IS NOT NULL'
-        )
-        .get(videoId)
-  return (row as { n: number }).n
-}
 
 function listVideoAssetPaths(videoId: number, type: string): Array<string | null> {
   return getDb()
@@ -756,90 +645,6 @@ export function getVideoImageCandidatePaths(videoId: number): {
   return { coverPath, samplePaths: listVideoAssetPaths(videoId, 'sample') }
 }
 
-function isVideoFieldEmptyForFill(
-  video: Video,
-  field: VideoScrapeField,
-  femaleCastCount: number,
-  maleCastCount: number,
-  tagCount: number,
-  sourceName?: string,
-  ratingSourceName?: string,
-  imageFacts: VideoImageAvailabilityFacts = {}
-): boolean {
-  switch (field) {
-    case 'title':
-      return isBlankText(video.title)
-    case 'summary':
-      return isBlankText(video.summary)
-    case 'cover':
-      return !(imageFacts.coverAvailable ?? Boolean(video.cover_path?.trim()))
-    case 'releaseDate':
-      return isBlankText(video.release_date)
-    case 'maker':
-      return isBlankText(video.maker)
-    case 'publisher':
-      return isBlankText(video.publisher)
-    case 'series':
-      return isBlankText(video.series)
-    case 'director':
-      return isBlankText(video.director)
-    case 'duration':
-      return video.duration_seconds == null
-    case 'actressesFemale':
-      return femaleCastCount === 0
-    case 'actressesMale':
-      return maleCastCount === 0
-    case 'tags':
-      return tagCount === 0
-    case 'source':
-      return countVideoExternalIds(video.id, sourceName) === 0
-    case 'rating':
-      return countVideoExternalStats(video.id, ratingSourceName ?? sourceName) === 0
-    case 'samples': {
-      const paths = listVideoAssetPaths(video.id, 'sample')
-      return paths.length === 0 || !(imageFacts.samplePathsAvailable ?? paths.every(Boolean))
-    }
-    default:
-      return false
-  }
-}
-
-/**
- * In fillEmpty mode, keep only selected fields that are currently empty on the video.
- * Cast fields apply only when the video has no linked performers (no female and no male).
- */
-export function resolveEffectiveScrapeFields(
-  videoId: number,
-  fields: VideoScrapeField[],
-  mode: VideoScrapeUpdateMode = 'replace',
-  sourceName?: string,
-  ratingSourceName?: string,
-  imageFacts: VideoImageAvailabilityFacts = {}
-): VideoScrapeField[] {
-  if (mode !== 'fillEmpty') return fields
-  const video = getVideoById(videoId)
-  if (!video) return []
-  const requested = new Set(fields)
-  const femaleCastCount = requested.has('actressesFemale')
-    ? countVideoCastByGender(videoId, 'female')
-    : 0
-  const maleCastCount = requested.has('actressesMale')
-    ? countVideoCastByGender(videoId, 'male')
-    : 0
-  const tagCount = requested.has('tags') ? countScrapedVideoTags(videoId) : 0
-  return fields.filter((field) =>
-    isVideoFieldEmptyForFill(
-      video,
-      field,
-      femaleCastCount,
-      maleCastCount,
-      tagCount,
-      sourceName,
-      ratingSourceName,
-      imageFacts
-    )
-  )
-}
 
 /** Replace a video's cast from manual edit lists. */
 function replaceVideoCast(
@@ -1058,21 +863,9 @@ export function listVideosForBatchScrape(
 ): VideoBatchTarget[] {
   const db = getDb()
   const { sql: where, params } = buildBatchScrapeWhere(filter)
-  const candidates = db
+  return db
     .prepare(`SELECT v.id, v.code FROM videos v ${where} ORDER BY v.add_time`)
     .all(...params) as VideoBatchTarget[]
-  const missingFields = Array.from(new Set(filter.missingFields ?? []))
-  if (missingFields.length === 0) return candidates
-  return candidates.filter(
-    (video) =>
-      resolveEffectiveScrapeFields(
-        video.id,
-        missingFields,
-        'fillEmpty',
-        filter.sourceName,
-        filter.ratingSourceName
-      ).length > 0
-  )
 }
 
 export function countVideosForBatchScrape(
@@ -1107,358 +900,6 @@ export function markScrapeSucceeded(id: number): void {
   ).run({ id, scrapedAt })
 }
 
-export type VideoScrapeImpactAction = 'preserve' | 'set' | 'replace' | 'clear'
-export type VideoScrapeImpactReason =
-  | 'replace'
-  | 'fillEmpty'
-  | 'replaceIfPresent'
-  | 'existingValue'
-  | 'noValue'
-  | 'resourceUnavailable'
-
-export interface VideoScrapeFieldImpact {
-  field: VideoScrapeField
-  action: VideoScrapeImpactAction
-  reason: VideoScrapeImpactReason
-  currentValue: unknown
-  nextValue: unknown
-  sourceName?: string
-}
-
-export interface VideoScrapeApplicationPlan {
-  effectiveFields: VideoScrapeField[]
-  impacts: VideoScrapeFieldImpact[]
-  shouldApply: boolean
-  warnings: string[]
-}
-
-function normalizedScrapeText(value: string | null | undefined): string | null {
-  const normalized = value?.trim()
-  return normalized ? normalized : null
-}
-
-/** Build a read-only field plan before any database row or asset reference is changed. */
-export function planVideoScrapeResult(
-  videoId: number,
-  result: ScrapeResult,
-  coverRelPath: string | null,
-  sampleRelPaths: Array<string | null> = [],
-  fields?: VideoScrapeField[],
-  sourceName?: string,
-  mode: VideoScrapeUpdateMode = 'replace',
-  ratingSourceName?: string,
-  imageFacts: VideoImageAvailabilityFacts = {}
-): VideoScrapeApplicationPlan {
-  const requested = fields ?? ALL_VIDEO_SCRAPE_FIELDS
-  const effectiveFields = resolveEffectiveScrapeFields(
-    videoId,
-    requested,
-    mode,
-    sourceName,
-    ratingSourceName,
-    imageFacts
-  )
-  const effective = new Set(effectiveFields)
-  const video = getVideoById(videoId)
-  const warnings: string[] = []
-  const cast = result.actresses ?? []
-  const sampleUrls = (result.sampleImageUrls ?? []).filter((url) =>
-    Boolean(normalizedScrapeText(url))
-  )
-  const samplesReady =
-    sampleUrls.length > 0 &&
-    sampleRelPaths.length === sampleUrls.length &&
-    sampleRelPaths.every((assetPath) => Boolean(assetPath))
-  const coverUnavailable =
-    effective.has('cover') && Boolean(normalizedScrapeText(result.coverUrl)) && !coverRelPath
-  const samplesUnavailable = effective.has('samples') && sampleUrls.length > 0 && !samplesReady
-  if (coverUnavailable) warnings.push('封面下载失败，已保留原封面')
-  if (samplesUnavailable) warnings.push('样张下载不完整，已保留原样张')
-
-  const durationValue =
-    typeof result.durationSeconds === 'number' &&
-    Number.isFinite(result.durationSeconds) &&
-    result.durationSeconds >= 0
-      ? result.durationSeconds
-      : null
-  const ratingValue =
-    typeof result.ratingAverage === 'number' && Number.isFinite(result.ratingAverage)
-      ? result.ratingAverage
-      : null
-  const values = new Map<VideoScrapeField, unknown>([
-    ['title', normalizedScrapeText(result.title)],
-    ['summary', normalizedScrapeText(result.summary)],
-    ['cover', coverRelPath],
-    ['releaseDate', normalizedScrapeText(result.releaseDate)],
-    ['maker', normalizedScrapeText(result.maker)],
-    ['publisher', normalizedScrapeText(result.publisher)],
-    ['series', normalizedScrapeText(result.series)],
-    ['director', normalizedScrapeText(result.director)],
-    ['duration', durationValue],
-    ['actressesFemale', cast.filter((item) => scrapedCastGender(item) === 'female')],
-    ['actressesMale', cast.filter((item) => scrapedCastGender(item) === 'male')],
-    ['tags', result.tags ?? []],
-    ['source', normalizedScrapeText(result.sourceUrl)],
-    ['rating', ratingValue],
-    ['samples', samplesReady ? sampleRelPaths : []]
-  ])
-  const currentValues = new Map<VideoScrapeField, unknown>([
-    ['title', video?.title ?? null],
-    ['summary', video?.summary ?? null],
-    ['cover', video?.cover_path ?? null],
-    ['releaseDate', video?.release_date ?? null],
-    ['maker', video?.maker ?? null],
-    ['publisher', video?.publisher ?? null],
-    ['series', video?.series ?? null],
-    ['director', video?.director ?? null],
-    ['duration', video?.duration_seconds ?? null],
-    ['actressesFemale', countVideoCastByGender(videoId, 'female')],
-    ['actressesMale', countVideoCastByGender(videoId, 'male')],
-    ['tags', countScrapedVideoTags(videoId)],
-    ['source', countVideoExternalIds(videoId, sourceName)],
-    ['rating', countVideoExternalStats(videoId, ratingSourceName ?? sourceName)],
-    ['samples', listVideoAssetPaths(videoId, 'sample')]
-  ])
-
-  const impacts = requested.map((field): VideoScrapeFieldImpact => {
-    const nextValue = values.get(field) ?? null
-    const isCollection = Array.isArray(nextValue)
-    const hasValue = isCollection ? nextValue.length > 0 : nextValue !== null && nextValue !== undefined
-    const resourceUnavailable =
-      (field === 'cover' && coverUnavailable) || (field === 'samples' && samplesUnavailable)
-    const missingFieldSource =
-      (field === 'source' && !sourceName) ||
-      (field === 'rating' && !(ratingSourceName ?? sourceName))
-    let action: VideoScrapeImpactAction = 'preserve'
-    let reason: VideoScrapeImpactReason = effective.has(field) ? 'noValue' : 'existingValue'
-    if (effective.has(field) && missingFieldSource) {
-      reason = 'noValue'
-    } else if (effective.has(field) && resourceUnavailable) {
-      reason = 'resourceUnavailable'
-    } else if (effective.has(field) && mode === 'replace') {
-      action = hasValue ? 'replace' : 'clear'
-      reason = 'replace'
-    } else if (effective.has(field) && hasValue) {
-      action = mode === 'fillEmpty' ? 'set' : 'replace'
-      reason = mode
-    }
-    return {
-      field,
-      action,
-      reason,
-      currentValue: currentValues.get(field) ?? null,
-      nextValue,
-      sourceName:
-        field === 'rating'
-          ? (ratingSourceName ?? sourceName)
-          : field === 'source'
-            ? sourceName
-            : undefined
-    }
-  })
-
-  return {
-    effectiveFields,
-    impacts,
-    shouldApply: impacts.some((impact) => impact.action !== 'preserve'),
-    warnings
-  }
-}
-
-/**
- * Apply a scrape result to a video, link actresses/tags, and store cover path.
- * Wrapped in a transaction so a partial failure doesn't corrupt relations.
- */
-export interface ApplyVideoScrapeResult {
-  applied: boolean
-  warnings: string[]
-  obsoleteAssetPaths: string[]
-}
-
-export function applyScrapeResult(
-  videoId: number,
-  result: ScrapeResult,
-  coverRelPath: string | null,
-  actressAvatars: Map<string, string | null>,
-  sampleRelPaths: Array<string | null> = [],
-  fields?: VideoScrapeField[],
-  sourceName?: string,
-  mode: VideoScrapeUpdateMode = 'replace',
-  ratingSourceName?: string,
-  imageFacts: VideoImageAvailabilityFacts = {}
-): ApplyVideoScrapeResult {
-  const db = getDb()
-  const requested = fields ?? ALL_VIDEO_SCRAPE_FIELDS
-  const plan = planVideoScrapeResult(
-    videoId,
-    result,
-    coverRelPath,
-    sampleRelPaths,
-    requested,
-    sourceName,
-    mode,
-    ratingSourceName,
-    imageFacts
-  )
-  if (!plan.shouldApply) return { applied: false, warnings: plan.warnings, obsoleteAssetPaths: [] }
-
-  const impactFor = (field: VideoScrapeField): VideoScrapeFieldImpact | undefined =>
-    plan.impacts.find((impact) => impact.field === field)
-  const impactAction = (field: VideoScrapeField): VideoScrapeImpactAction =>
-    impactFor(field)?.action ?? 'preserve'
-  const writesField = (field: VideoScrapeField): boolean => impactAction(field) !== 'preserve'
-  const existing = getVideoById(videoId)
-  if (!existing) return { applied: false, warnings: [], obsoleteAssetPaths: [] }
-  const cleanupHints = collectVideoLibraryCleanupHints(videoId)
-  const scrapedAt = nowIso()
-  const warnings = plan.warnings
-  const scalarColumns: Partial<Record<VideoScrapeField, { column: string; bindKey: string }>> = {
-    title: { column: 'title', bindKey: 'title' },
-    summary: { column: 'summary', bindKey: 'summary' },
-    releaseDate: { column: 'release_date', bindKey: 'release_date' },
-    maker: { column: 'maker', bindKey: 'maker' },
-    publisher: { column: 'publisher', bindKey: 'publisher' },
-    series: { column: 'series', bindKey: 'series' },
-    director: { column: 'director', bindKey: 'director' },
-    duration: { column: 'duration_seconds', bindKey: 'duration_seconds' }
-  }
-  const scalarWrites = plan.impacts.flatMap((impact) => {
-    const metadata = scalarColumns[impact.field]
-    return metadata && impact.action !== 'preserve'
-      ? [{ field: impact.field, value: impact.nextValue, ...metadata }]
-      : []
-  })
-
-  const cast = result.actresses ?? []
-  const writeFemale = writesField('actressesFemale')
-  const writeMale = writesField('actressesMale')
-  const writeTags = writesField('tags')
-
-  const sourceUrl = (impactFor('source')?.nextValue as string | null | undefined) ?? null
-  const writeSource = Boolean(sourceName && writesField('source'))
-  const ratingValue = (impactFor('rating')?.nextValue as number | null | undefined) ?? null
-  const statsSource = ratingSourceName ?? sourceName
-  const writeRating = Boolean(statsSource && writesField('rating'))
-
-  const writeCover = writesField('cover')
-
-  const sampleUrls = (result.sampleImageUrls ?? []).filter((url) =>
-    Boolean(normalizedScrapeText(url))
-  )
-  const writeSamples = writesField('samples')
-
-  const oldAssetPaths: Array<string | null> = []
-
-  const txn = db.transaction(() => {
-    const assignments: string[] = []
-    const bind: Record<string, unknown> = { id: videoId }
-
-    for (const { field, column, bindKey, value } of scalarWrites) {
-      assignments.push(`${column} = @${bindKey}`)
-      bind[bindKey] = value
-      if (field === 'title') {
-        assignments.push('original_title = @original_title')
-        bind.original_title = value
-      }
-    }
-
-    if (writeCover) {
-      assignments.push('cover_path = @cover_path')
-      bind.cover_path = coverRelPath
-      oldAssetPaths.push(existing.cover_path)
-      oldAssetPaths.push(
-        ...replaceVideoAssets(
-          videoId,
-          'cover',
-          coverRelPath
-            ? [
-                {
-                  type: 'cover',
-                  position: 0,
-                  remoteUrl: result.coverUrl ?? null,
-                  localPath: coverRelPath,
-                  isPrimary: 1,
-                  createdAt: scrapedAt
-                }
-              ]
-            : []
-        )
-      )
-    }
-
-    if (assignments.length > 0) {
-      db.prepare(`UPDATE videos SET ${assignments.join(', ')} WHERE id = @id`).run(bind)
-    }
-
-    if (writeFemale) {
-      removeVideoActressesByGender(videoId, 'female')
-      linkScrapedCastByGender(videoId, cast, 'female', actressAvatars)
-    }
-    if (writeMale) {
-      removeVideoActressesByGender(videoId, 'male')
-      linkScrapedCastByGender(videoId, cast, 'male', actressAvatars)
-    }
-
-    if (writeTags) {
-      replaceTagsByOrigin(videoId, result.tags ?? [], 'scraped', sourceName ?? null, scrapedAt)
-    }
-
-    ensureFacetEntries({
-      maker: writesField('maker') ? (result.maker ?? null) : null,
-      publisher: writesField('publisher') ? (result.publisher ?? null) : null,
-      series: writesField('series') ? (result.series ?? null) : null,
-      director: writesField('director') ? (result.director ?? null) : null
-    })
-
-    if (writeSamples) {
-      oldAssetPaths.push(
-        ...replaceVideoAssets(
-          videoId,
-          'sample',
-          sampleUrls.map((url, index) => ({
-            type: 'sample',
-            position: index,
-            remoteUrl: url,
-            localPath: sampleRelPaths[index]!,
-            isPrimary: 0,
-            createdAt: scrapedAt
-          }))
-        )
-      )
-    }
-
-    if (writeSource && sourceName) {
-      if (sourceUrl) upsertVideoExternalId(videoId, sourceName, result, scrapedAt)
-      else deleteVideoExternalId(videoId, sourceName)
-    }
-
-    if (writeRating && statsSource) {
-      if (ratingValue !== null) {
-        upsertVideoExternalStats(videoId, statsSource, result, scrapedAt)
-      } else {
-        deleteVideoExternalStats(videoId, statsSource)
-      }
-    }
-
-    db.prepare(
-      'UPDATE videos SET scraped_status = 1, last_scraped_at = ?, updated_at = ? WHERE id = ?'
-    ).run(scrapedAt, scrapedAt, videoId)
-  })
-
-  txn()
-
-  const obsoleteAssetPaths = Array.from(new Set(oldAssetPaths)).filter(
-    (assetPath): assetPath is string =>
-      Boolean(assetPath && assetPath !== coverRelPath && !sampleRelPaths.includes(assetPath))
-  )
-
-  try {
-    runLibraryCleanup(cleanupHints)
-  } catch (error) {
-    console.error('Post-commit library cleanup failed:', error)
-  }
-  return { applied: true, warnings, obsoleteAssetPaths }
-}
 
 /** Update stored relative asset paths after encrypt/decrypt migration. */
 export function remapAssetPath(oldRel: string, newRel: string): void {
@@ -1479,121 +920,7 @@ export function remapAssetPath(oldRel: string, newRel: string): void {
   )
 }
 
-function upsertVideoAsset(
-  videoId: number,
-  asset: {
-    type: string
-    position: number
-    remoteUrl: string | null
-    localPath: string | null
-    isPrimary: number
-    createdAt: string
-  }
-): void {
-  const db = getDb()
-  if (asset.isPrimary) {
-    db.prepare('UPDATE video_assets SET is_primary = 0 WHERE video_id = ? AND type = ?').run(
-      videoId,
-      asset.type
-    )
-  }
-  db.prepare(
-    `INSERT INTO video_assets
-       (video_id, type, position, remote_url, local_path, is_primary, created_at)
-     VALUES (@videoId, @type, @position, @remoteUrl, @localPath, @isPrimary, @createdAt)`
-  ).run({ videoId, ...asset })
-}
 
-function replaceVideoAssets(
-  videoId: number,
-  type: string,
-  assets: Array<{
-    type: string
-    position: number
-    remoteUrl: string | null
-    localPath: string | null
-    isPrimary: number
-    createdAt: string
-  }>
-): Array<string | null> {
-  const db = getDb()
-  const old = db
-    .prepare('SELECT local_path FROM video_assets WHERE video_id = ? AND type = ?')
-    .all(videoId, type) as { local_path: string | null }[]
-  clearVideoPosterForPaths(videoId, old.map((row) => row.local_path))
-  db.prepare('DELETE FROM video_assets WHERE video_id = ? AND type = ?').run(videoId, type)
-  for (const asset of assets) {
-    upsertVideoAsset(videoId, asset)
-  }
-  return old.map((row) => row.local_path)
-}
-
-function upsertVideoExternalId(
-  videoId: number,
-  source: string,
-  result: ScrapeResult,
-  fetchedAt: string
-): void {
-  const db = getDb()
-  db.prepare(
-    `INSERT INTO video_external_ids
-       (video_id, source, external_id, external_code, url, title, fetched_at)
-     VALUES (@videoId, @source, @externalId, @externalCode, @url, @title, @fetchedAt)
-     ON CONFLICT(video_id, source) DO UPDATE SET
-       external_id = excluded.external_id,
-       external_code = excluded.external_code,
-       url = excluded.url,
-       title = excluded.title,
-       fetched_at = excluded.fetched_at`
-  ).run({
-    videoId,
-    source,
-    externalId: null,
-    externalCode: result.code || null,
-    url: result.sourceUrl ?? null,
-    title: result.title ?? null,
-    fetchedAt
-  })
-}
-
-function deleteVideoExternalId(videoId: number, source: string): void {
-  getDb()
-    .prepare('DELETE FROM video_external_ids WHERE video_id = ? AND source = ?')
-    .run(videoId, source)
-}
-
-function upsertVideoExternalStats(
-  videoId: number,
-  source: string,
-  result: ScrapeResult,
-  fetchedAt: string
-): void {
-  if (result.ratingAverage === undefined && result.ratingCount === undefined) return
-  const db = getDb()
-  db.prepare(
-    `INSERT INTO video_external_stats
-       (video_id, source, rating_average, rating_count, fetched_at)
-     VALUES (@videoId, @source, @ratingAverage, @ratingCount, @fetchedAt)
-     ON CONFLICT(video_id, source) DO UPDATE SET
-       rating_average = excluded.rating_average,
-       rating_count = excluded.rating_count,
-       fetched_at = excluded.fetched_at`
-  ).run({
-    videoId,
-    source,
-    ratingAverage: result.ratingAverage ?? null,
-    ratingCount: result.ratingCount ?? null,
-    fetchedAt
-  })
-}
-
-function deleteVideoExternalStats(videoId: number, source: string): void {
-  const db = getDb()
-  db.prepare('DELETE FROM video_external_stats WHERE video_id = ? AND source = ?').run(
-    videoId,
-    source
-  )
-}
 
 function deleteVideoAssetRows(videoId: number): string[] {
   const db = getDb()
